+++ BambooTracker/gui/pattern_editor/pattern_editor.hpp
@@ -40,6 +40,7 @@ signals:
 	void selected(bool isSelected);
 	void returnPressed();
 	void instrumentEntered(int num);
+	void effectEntered(QString text);
 
 protected:
 	bool eventFilter(QObject *watched, QEvent *event) override;