+++ BambooTracker/gui/pattern_editor/pattern_editor_panel.hpp
@@ -88,6 +88,7 @@ signals:
 	void selected(bool isSelected);
 	void returnPressed();
 	void instrumentEntered(int num);
+	void effectEntered(QString text);
 
 protected:
 	virtual bool event(QEvent *event) override;